requirement.txt
flask
flask_sqlalchemy
flask_mysqldb
flask_script
    from ._compat import text_type 改成 from flask_script._compat import text_type
flask_debugtoolbar
pymysql
apscheduler
flask-apscheduler
requests
beautifulsoup4
flask-sqlacodegen
uwsgi